requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
azure-storage-blob
python-dotenv